pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "sunwhere"
version = "1.3.4"
description = "Calculation of solar geometry for solar resource assessment applications"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Jose A Ruiz-Arias", email = "jararias@uma.es"}]
license = {text = "CC BY-NC-SA 4.0"}
classifiers = [
    "Natural Language :: English",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "loguru>=0.7.3",
    "numexpr>=2.10.0",
    "numpy>=2.0.0",
    "pandas>=2.2.2",
    "xarray>=2025.6.1",
]

[project.urls]
Homepage = "https://github.com/jararias/sunwhere"

[dependency-groups]
# estas son las dependencias que se supone que solo los desarrolladores van a
# querer (posiblemente) instalar. Al menos el grupo dev se instala por defecto
# por uv cuando trabajamos en modo "editable"
# con uv hay que añadirlos como: uv add --dev <package1> <package 2> ...
dev = [
    "coverage>=7.13.5",
    "genbadge[tests]>=1.1.3",
    "ipython>=8.39.0",
    "lazydocs",
    "pvlib>=0.15.1",
    "pytest>=9.0.3",
    "pytest-allclose>=1.0.0",
    "ruff>=0.15.12",
    "seaborn>=0.13.2",
    "soltrack",
    "tqdm>=4.67.3",
    "urllib3>=2.6.3",
]

[project.optional-dependencies]
# estas son dependencias que un usuario final de la libreria puede querer instalar.
# son las que se instalan como pip install libreria[opcion]
# con uv hay que añadirlas como: uv add --optional cli <package1> <package2> ...
cli = [
    "matplotlib>=3.10.9",
    "python-dateutil>=2.9.0.post0",
    "tabulate>=0.10.0",
    "tqdm>=4.67.3",
    "typer>=0.25.0",
]

[tool.uv]
# Esto asegura que uv sepa cómo resolver la dependencia de Git
managed = true

[tool.uv.sources]
soltrack = { git = "https://github.com/jararias/SolTrack" }
lazydocs = { git = "https://github.com/jararias/lazydocs" }

[project.scripts]
sunwhere = "sunwhere._cli.main:app"

[tool.coverage.run]
omit = ["sunwhere/utils/*"]
